requirements.txt
ultralytics
pandas
seaborn
matplotlib
opencv-python
scipy
tqdm
pyinstaller
moviepy
pupil-labs-real-time-api